+++ app/src/main/java/com/app/arcabyolimpo/presentation/navigation/NavGraph.kt
@@ -1,5 +1,6 @@
 package com.app.arcabyolimpo.presentation.navigation
 
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
@@ -15,6 +16,7 @@ import androidx.navigation.navArgument
 import androidx.navigation.navDeepLink
 import com.app.arcabyolimpo.data.remote.interceptor.SessionManager
 import com.app.arcabyolimpo.domain.model.auth.UserRole
+import com.app.arcabyolimpo.presentation.common.components.LoadingShimmer
 import com.app.arcabyolimpo.presentation.screens.accountactivation.AccountActivationScreen
 import com.app.arcabyolimpo.presentation.screens.home.assistant.CollaboratorHomeScreen
 import com.app.arcabyolimpo.presentation.screens.home.coordinator.CoordinatorHomeScreen
@@ -214,15 +216,22 @@ fun ArcaNavGraph(
                     )
                 }
 
-                uiState.response?.valid == false -> {
-                    TokenVerificationFailedScreen(onBackClick = { navController.popBackStack() })
+                uiState.isLoading -> {
+                    LoadingShimmer(
+                        modifier = Modifier
+                            .fillMaxSize()
+                    )
                 }
 
-                uiState.isLoading == true -> {
+                uiState.error == "Invalid or expired token" -> {
+                    TokenVerificationFailedScreen(onBackClick = { navController.popBackStack() })
                 }
 
                 else -> {
-                    TokenVerificationFailedScreen(onBackClick = { navController.popBackStack() })
+                    LoadingShimmer(
+                        modifier = Modifier
+                            .fillMaxSize()
+                    )
                 }
             }
         }